Added CLI support for input and output files + implemented vertical sorting as well as its CLI support

+++ include/hbs.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-int get_hue(char r, char g, char b);
-int get_saturation(char r, char g, char b);
-int get_brightness(char r, char g, char b);
+float get_hue(char r, char g, char b);
+float get_saturation(char r, char g, char b);
+float get_brightness(char r, char g, char b);
